+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.js
@@ -416,7 +416,8 @@ export let SidePanel = React.forwardRef(
               onAnimationStart={onAnimationStart}
               onBlur={handleBlur}
               ref={ref || sidePanelRef}
-              role="complementary">
+              role="complementary"
+              aria-label={title}>
               <span
                 ref={startTrapRef}
                 tabIndex="0"
@@ -461,6 +462,7 @@ export let SidePanel = React.forwardRef(
                     </h5>
                   )}
                   <Button
+                    hasIconOnly
                     kind="ghost"
                     size="small"
                     disabled={false}
@@ -476,7 +478,8 @@ export let SidePanel = React.forwardRef(
                 {subtitle && subtitle.length && (
                   <p
                     className={cx(`${blockClass}__subtitle-text`, {
-                      [`${blockClass}__subtitle-text-no-animation`]: !animateTitle,
+                      [`${blockClass}__subtitle-text-no-animation`]:
+                        !animateTitle,
                       [`${blockClass}__subtitle-text-no-animation-no-action-toolbar`]:
                         !animateTitle &&
                         (!actionToolbarButtons || !actionToolbarButtons.length),
@@ -487,7 +490,8 @@ export let SidePanel = React.forwardRef(
                 {actionToolbarButtons && actionToolbarButtons.length && (
                   <div
                     className={cx(`${blockClass}__action-toolbar`, {
-                      [`${blockClass}__action-toolbar-no-animation`]: !animateTitle,
+                      [`${blockClass}__action-toolbar-no-animation`]:
+                        !animateTitle,
                     })}>
                     {actionToolbarButtons.map((action) => (
                       <Button
@@ -502,8 +506,10 @@ export let SidePanel = React.forwardRef(
                         className={cx([
                           `${blockClass}__action-toolbar-button`,
                           {
-                            [`${blockClass}__action-toolbar-icon-only-button`]: action.icon,
-                            [`${blockClass}__action-toolbar-leading-button`]: !action.icon,
+                            [`${blockClass}__action-toolbar-icon-only-button`]:
+                              action.icon,
+                            [`${blockClass}__action-toolbar-leading-button`]:
+                              !action.icon,
                           },
                         ])}
                         onClick={() => action.onActionToolbarButtonClick()}>
@@ -710,7 +716,7 @@ SidePanel.propTypes = {
   /**
    * Sets the title text
    */
-  title: PropTypes.string,
+  title: PropTypes.string.isRequired,
 };
 
 SidePanel.defaultProps = {