+++ packages/jest-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.10](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/jest-config@0.3.9...jest-config@0.3.10) (2021-05-28)
+
+**Note:** Version bump only for package jest-config
+
+
+
+
+
 ## [0.3.9](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/jest-config@0.3.8...jest-config@0.3.9) (2021-05-26)
 
 **Note:** Version bump only for package jest-config